+++ packages/mermaid/src/diagrams/flowchart/types.ts
@@ -11,6 +11,12 @@ export interface FlowVertex {
   styles: string[];
   text?: string;
   type?: string;
+  icon?: string;
+  form?: string;
+  pos?: 't' | 'b';
+  img?: string;
+  assetWidth?: number;
+  assetHeight?: number;
 }
 
 export interface FlowText {