+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/subroutine.ts
@@ -1,13 +1,10 @@
 import { labelHelper, updateNodeBounds, getNodeClasses } from './util.js';
 import intersect from '../intersect/index.js';
-import type { Node } from '$root/rendering-util/types.d.ts';
-import { getConfig } from '$root/diagram-api/diagramAPI.js';
-import {
-  styles2String,
-  userNodeOverrides,
-} from '$root/rendering-util/rendering-elements/shapes/handDrawnShapeStyles.js';
+import type { Node } from '../../types.js';
+import { styles2String, userNodeOverrides } from './handDrawnShapeStyles.js';
 import rough from 'roughjs';
 import { insertPolygonShape } from './insertPolygonShape.js';
+import { getConfig } from '../../../config.js';
 
 export const createSubroutinePathD = (
   x: number,