+++ packages/mermaid/src/rendering-util/rendering-elements/shapes/util.js
@@ -1,8 +1,8 @@
-import { createText } from '$root/rendering-util/createText.ts';
-import { getConfig } from '$root/diagram-api/diagramAPI.js';
+import { createText } from '../../createText.js';
+import { getConfig } from '../../../diagram-api/diagramAPI.js';
 import { select } from 'd3';
-import { evaluate, sanitizeText } from '$root/diagrams/common/common.js';
-import { decodeEntities } from '$root/utils.js';
+import { evaluate, sanitizeText } from '../../../diagrams/common/common.js';
+import { decodeEntities } from '../../../utils.js';
 
 export const labelHelper = async (parent, node, _classes) => {
   let cssClasses;
@@ -36,6 +36,7 @@ export const labelHelper = async (parent, node, _classes) => {
     width: node.width || getConfig().flowchart.wrappingWidth,
     cssClasses: 'markdown-node-label',
     style: node.labelStyle,
+    addSvgBackground: !!node.icon || !!node.img,
   });
   // Get the size of the label
   let bbox = text.getBBox();