+++ packages/mermaid/src/types.ts
@@ -1,7 +1,16 @@
 export interface NodeMetaData {
   shape?: string;
   label?: string;
+  icon?: string;
+  form?: string;
+  pos?: 't' | 'b';
+  img?: string;
+  w?: string;
+  h?: string;
 }
+import type { MermaidConfig } from './config.type.js';
+import type { Diagram } from './Diagram.ts';
+
 export interface Point {
   x: number;
   y: number;
@@ -52,6 +61,18 @@ export interface ParseResult {
    * The diagram type, e.g. 'flowchart', 'sequence', etc.
    */
   diagramType: string;
+  /**
+   * The config passed as YAML frontmatter or directives
+   */
+  config: MermaidConfig;
+
+  defaultConfig?: MermaidConfig;
+
+  /**
+   * The diagram AST
+   *
+   */
+  diagram: Diagram;
 }
 // This makes it clear that we're working with a d3 selected element of some kind, even though it's hard to specify the exact type.
 export type D3Element = any;